refactor complete! works like before.
