+++ src/components/ScarlettMixerChannel.vue
@@ -1,5 +1,5 @@
 <template>
-  <div id="channel1" >
+  <div>
     <div id="input"></div>
     <div id="pan"></div>
     <div id="volume"></div>
@@ -8,9 +8,6 @@
 </template>
 
 <script>
-  // element id's and names will come from the parent component
-  // the loadAudio() function usefulness has expired
-  // this will require moving all the UI logic into this component, with params
   export default {
     data () {
       return {
